Moving Bucketing related logic into Bucketing folder and namespace.  Also matches the Java SDK structure.
