+++ OptimizelySDK/Bucketing/UserProfileUtil.cs
@@ -17,7 +17,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace OptimizelySDK.Utils
+namespace OptimizelySDK.Bucketing
 {
     public class UserProfileUtils
     {
