+++ OptimizelySDK/OptimizelySDK.csproj
@@ -61,8 +61,8 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Bucketer.cs" />
-    <Compile Include="Decision.cs" />
+    <Compile Include="Bucketing\Bucketer.cs" />
+    <Compile Include="Bucketing\Decision.cs" />
     <Compile Include="Entity\Attribute.cs" />
     <Compile Include="Entity\Audience.cs" />
     <Compile Include="Entity\Entity.cs" />
@@ -91,16 +91,16 @@
     <Compile Include="Logger\NoOpLogger.cs" />
     <Compile Include="Optimizely.cs" />
     <Compile Include="ProjectConfig.cs" />
-    <Compile Include="UserProfile.cs" />
+    <Compile Include="Bucketing\UserProfile.cs" />
     <Compile Include="Utils\ConditionEvaluator.cs" />
     <Compile Include="Utils\EventTagUtils.cs" />
-    <Compile Include="Utils\UserProfileUtil.cs" />
+    <Compile Include="Bucketing\UserProfileUtil.cs" />
     <Compile Include="Utils\Validator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils\ConfigParser.cs" />
     <Compile Include="Utils\Schema.cs" />
-    <Compile Include="UserProfileService.cs" />
-    <Compile Include="DecisionService.cs" />
+    <Compile Include="Bucketing\UserProfileService.cs" />
+    <Compile Include="Bucketing\DecisionService.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Utils\schema.json" />
